+++ packages/mrm-task-lint-staged-simple-git-hooks/index.js
@@ -1,10 +1,15 @@
 // @ts-check
-const { packageJson, install, getExtsFromCommand } = require('mrm-core');
+const {
+	packageJson,
+	install,
+	getExtsFromCommand,
+	uninstall,
+} = require('mrm-core');
 const { castArray } = require('lodash');
 
 const packages = {
 	'lint-staged': '>=10',
-	husky: '=4',
+	'simple-git-hooks': '>=2.0.3',
 };
 
 /**
@@ -172,17 +177,18 @@ module.exports = function task({ lintStagedRules }) {
 	pkg
 		// Remove husky 0.14 config
 		.unset('scripts.precommit')
+		// Remove husky 4 config
+		.unset('husky')
 		// Add new config
 		.merge({
-			husky: {
-				hooks: {
-					'pre-commit': 'lint-staged',
-				},
+			'simple-git-hooks': {
+				'pre-commit': 'npx lint-staged',
 			},
 			'lint-staged': rules,
 		})
 		.save();
 
+	uninstall('husky');
 	// Install dependencies
 	install(packages);
 };